Home page refactor
Implement new asset routes

Refactor code for quality and formatting

Clean components and finalize
