+++ components/HomePage/GithubCard.tsx
@@ -5,7 +5,7 @@ import React, { useContext, useEffect, useState } from "react";
 import GitHubCalendar, { Theme } from "react-github-calendar";
 import { AppContext } from "../../Utils/AppContext";
 
-const css = require("./TechnicalSkills.module.css");
+const css = require("./Home.module.css");
 
 const calenderStyle = {
   width: "min(90%, 1400px)",
