+++ components/HomePage/TechnicalSkill.tsx
@@ -4,7 +4,7 @@ import React from "react";
 
 import { IconBaseProps } from "react-icons";
 
-const css = require("./TechnicalSkills.module.css");
+const css = require("./Home.module.css");
 
 interface skillProps {
   name: string;
