+++ pages/index.tsx
@@ -17,17 +17,18 @@ const Languages = dynamic(() => import("../components/HomePage/Languages"));
 const Framework = dynamic(() => import("../components/HomePage/Framework"));
 const Databases = dynamic(() => import("../components/HomePage/Databases"));
 
-const css = require("../components/HomePage/HomeTop.module.css");
+const css = require("../components/HomePage/Home.module.css");
 
 const HomePage = () => {
   const { windowWidth } = useContext(AppContext);
 
+  // Window scrolling affect on home page Title
   useEffect(() => {
     const animateScroll = () => {
-      const mobile = windowWidth == WindowWidth.SMALL;
+      const mobile = windowWidth === WindowWidth.SMALL;
       const slide = document.getElementById("slideTitle")!;
-      const down = document.getElementById("titleDown")!;
       const rect = slide.getBoundingClientRect();
+      const down = document.getElementById("titleDown")!;
 
       if (
         (rect.bottom <= window.innerHeight && rect.bottom >= 0) ||
@@ -57,6 +58,7 @@ const HomePage = () => {
         down.style.visibility = "hidden";
       }
     };
+
     window.addEventListener("scroll", animateScroll);
     return () => {
       window.removeEventListener("scroll", animateScroll);
@@ -67,7 +69,7 @@ const HomePage = () => {
     <>
       <MetaHead
         title="Jesse Greenough's Development Portfolio"
-        description="View the Development Portfolio of Full Stack Engineer Jesse Greenough"
+        description="View the Technical Development Skills possessed by Full Stack Engineer Jesse Greenough"
         keywords="Software,Developer,Engineer,Full-Stack,Portfolio,Skills,Projects,Experience,Resumes"
       />
 
@@ -76,7 +78,10 @@ const HomePage = () => {
         style={{ padding: "clamp(160px, 14vw, 240px) 0 0" }}
       >
         <HomeTop />
-        <section id="skillsIntro" className={css.developerSkills}>
+        <section
+          id="skillsIntro"
+          className={css.developerSkills}
+        >
           <h2 id="slideTitle">
             Check out my <span className="detail">Technical Skillsets</span>
             <br />
